auth: add central caching authentication source
The source of our authentication with GitHub is a private key that we use
to sign a json web authentication token and log in as GitHub App.
This is used directly when we want to act on behalf of the Forrest
application itself (to e.g. get the list of installations) and to request
short-lived installation token whenever we want to act on behalf of an
user.