config: implement config file reading and automatic re-reading
Restarting the Forrest service just to re-read the config file would be
quite annoying, because it would mean stopping all running jobs.
Instead we re-read the config file whenever it changed on disk.
If that fails we just keep the old version and log the issue.

A config that changes while the service is running could lead to
inconsistencies, like allocating X amount of RAM for a VM but starting
it with Y amount of ram because the config changed in the meantime.
We need to take care to avoid situations like these.