ingres: webhook: get job status updates via webhook events
Webhooks allow us to receive job status updates in timely manner
(at least quicker than polling).
To receive webhook events we do however need to provide a web server that
can be reached from the internet.
To keep the impact (in terms of security and code size) small we only
provide a super basic web server that is only able to handle webhook POST
requests and that does not even listen on a TCP port itself, using a Unix
domain socket instead, to strongly guide the user towards using a real
reverse proxy with TLS support etc. in front of it.

The server is hand-crafted because I was not satisfied with the first
few tokio HTTP servers I looked at (they were all rather heavy-weight
and had a lot of features we do not need) and because I always wanted to
implement a HTTP server.