jobs: manager: add a job manager that keeps track of job states
We receive job status updates by polling and via webhooks,
which means we need some central location where information is aggregated.
The job manager does just that. It keeps track of jobs and their state.

The job manager forwards the current demand for machines to the machine
manager, or it will once the machine manager is implemented.