machines: run_dir: set up a directory for each run
We need a filesystem directory per run where we can place the disk image,
config files, lock and an unix domain socket that allows interactive
debugging of the running machine.

This commit adds the basic directory setup and (reflink) copying the disk
image, but does not yet create the config files or the log/debug files,
which will be created by qemu instead.