machines: config_fs: set up filesystems for machine and job config
Configuration of the virtual machines is done via cloud-init[1] and the
job configuration is done via another filesystem,
containing a script to run.

Both cloud-init and job-config use FAT filesystem images that are created
via the `fatfs` crate and that are handed over to the virtual machines as
additional disk images.

The job can also provide feedback to Forrest by writing to files in the
job-config filesystem.
Currently this is only used to decide if the VM's disk image should be
persisted as new base image for this machine type.

[1]: https://cloud-init.io/