main: make the main function async
Our interaction with the GitHub API (polling and webhook) and spawning
virtual machines will be handled in an async fashion.
Prepare the `main` function accordingly.

We use a single thread to run everything because the load on Forrest
itself should be rather low.
Running the service multi-threaded could mask issues with the locking or
blocking code or the performance in general.