contrib: setup_templates: do not disable the ssh service
The cloud-init scripts should be kept as minimalistic as possible,
because they are (compared to jobs run via GitHub) hard to debug.
The ssh service is not reachable from the outside,
due to the NAT-ing done by qemu slirp/user networking.
We can thus leave the decision to disable/uninstall ssh to the setup code
in the jobs.