contrib: setup_templates: use write_file for root shell getty override
This results in the file being written unconditionally on every boot.
This also means that the file in cloud-init and the one in the machine
will not become out of sync.
The drawback is that we need to systemctl deamon-reload on every boot now,
which should however be okay performance-wise.