ingres: poll: do not poll before notifying systemd
We poll each user, repository, run and job in sequence.
This can take rather long when a lot of repositories are configured,
in fact too long for the default systemd job startup timeout.

There is no real benefit to delaying the startup notification until after
the first of polling, so just don't do it instead.
This speeds up the perceived startup a lot.