ingres: poll: only poll runs from the last 24 hours in initial poll
First of all: this does not affect runs that the Forrest instance already
knows about. This is only the initial scan for runs that need our
attention.

There can not be any runs older than 24 hours that need our attention,
because GitHub cancels jobs for which no runner has registered
after 24 hours.