ingres: webhook: use hyper instead of the ad-hoc http implementation
In the future we will likely want to do more things via HTTP,
namely serving prometheus metrics, providing a status page or
allowing the jobs inside of the machines to upload artifacts.

Growing a custom web server to do all of these things would be
possible, but not really a reasonable thing to do.

Instead use the hyper crate as a HTTP server implementation for the
webhook endpoint.

Using hyper as a server crate has the benefit of already being in our
dependencies via the octocrab crate.