Cargo,config: switch from serde_yml to serde_yaml_ng
The serde YAML situation is still complicated.
The serde_yml maintainer claims to maintain a whole lot of repositories
and to use AI to do so.
This leaves a bit of a bad taste and some people from the internet advice
against using his crates.
Play it safe and switch to a more conservative serde_yaml fork.