README: try to get to the point and don't focus on technical details
How images can be persisted between runs and how that can be used is
certainly interesting, but it is also a bit confusing and not what the
user *needs* to know when deciding if they want to use Forrest or not.
Try to get that across better and refer to the docs for details.