machines: triplets: rename from triplet
A triplet, in forrest speek, is the combination of a username, repository
and machine type.
This is about to change, because another kind of triplet will be added,
containing username, repository and labels.

Rename the module to indicate that it contains different kinds of
triplets. The old and now name are still not good though, because it is
very non-descriptive and because the file also contains e.g. the
OwnerAndRepo struct, which only has two components.

Leave proper naming to future us.