machines: triplets: add a OwnerRepoLabels struct
This is only slightly different from the `OwnerRepoMachine` struct.
It does not care if the contained Labels actually correspond to what we
expect in forrest (e.g. `runs-on: [self-hosted, forrest, machine-name]`)
and just preserves them as-is.

Why add this struct if `Triplet`/`OwnerRepoMachines` worked fine until
now?

We have to register runners with exactly the same set of labels as those
specified by the job and given to us via the API or webhooks, otherwise
the job will not be assigned to the runer.

This is much easier if we do not have to reproduce the labels from the
machine name and will allow us to accept the labels in different formats
in the future.