contrib: setup_templates: work around missing libicu on Debian trixie
The Debian Trixie cloud images do not contain a version of libicu,
which is required to run the GitHub action runner application.
Without it it will fail with the following error:

    Couldn't find a valid ICU package installed on the system.

Installing just libicu on Debian is however not that easy, because it
always comes with a version suffix like `libicu76` (trixie),
`libicu72` (bookworm) etc.

This is why we install `icu-devtools` instead, which depends on a
`libicu*` and is not too large itself.